Extract integration tests and add IntegrationTest target
